requirements.txt
streamlit
pandas
pillow
numpy
google-genai
requests